Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

NLPModelsTest.HS13

NLPModelsTest.check_nlp_dimensions

NLPModelsTest.check_nls_dimensions

NLPModelsTest.consistent_nlps

NLPModelsTest.consistent_nlss

NLPModelsTest.coord_memory_nlp

NLPModelsTest.gradient_check

NLPModelsTest.hessian_check

NLPModelsTest.hessian_check_from_grad

NLPModelsTest.jacobian_check

NLPModelsTest.multiple_precision_nlp

NLPModelsTest.multiple_precision_nls

NLPModelsTest.view_subarray_nlp

NLPModelsTest.view_subarray_nls

​

NLPModelsTest.HS13 — Type

nlp = HS13()

Problem 13 in the Hock-Schittkowski suite

\[\begin{aligned} \min \quad & (x_1 - 2)^2 + x_2^2 \\ \text{s. to} \quad & (1 - x_1)^3 - x_2 \geq 0 \end{aligned}\]

Starting point: [-2; -2].

NLPModelsTest.check_nlp_dimensions — Method

check_nlp_dimensions(nlp; exclude = [ghjvprod])

Make sure NLP API functions will throw DimensionError if the inputs are not the correct dimension. To make this assertion in your code use

@lencheck size input [more inputs separated by spaces]

NLPModelsTest.check_nls_dimensions — Method

check_nls_dimensions(nlp; exclude = [])

Make sure NLS API functions will throw DimensionError if the inputs are not the correct dimension. To make this assertion in your code use

@lencheck size input [more inputs separated by spaces]

NLPModelsTest.consistent_nlps — Method

consistent_nlps(nlps; exclude=[], rtol=1e-8)

Check that the all nlps of the vector nlps are consistent, in the sense that

Their counters are the same.

Their meta information is the same.

The API functions return the same output given the same input.

In other words, if you create two models of the same problem, they should be consistent.

The keyword exclude can be used to pass functions to be ignored, if some of the models don't implement that function.

NLPModelsTest.consistent_nlss — Method

consistent_nlss(nlps; exclude=[hess, hprod, hess_coord])

Check that the all nlss of the vector nlss are consistent, in the sense that

Their counters are the same.

Their meta information is the same.

The API functions return the same output given the same input.

In other words, if you create two models of the same problem, they should be consistent.

By default, the functions hess, hprod and hess_coord (and therefore associated functions) are excluded from this check, since some models don't implement them.

NLPModelsTest.coord_memory_nlp — Method

coord_memory_nlp(nlp; exclude = [])

Check that the allocated memory for in place coord methods is sufficiently smaller than their allocating counter parts.

NLPModelsTest.gradient_check — Method

gradient_check(nlp; x=nlp.meta.x0, atol=1e-6, rtol=1e-4)

Check the first derivatives of the objective at x against centered finite differences.

This function returns a dictionary indexed by components of the gradient for which the relative error exceeds rtol.

NLPModelsTest.hessian_check — Method

hessian_check(nlp; x=nlp.meta.x0, atol=1e-6, rtol=1e-4, sgn=1)

Check the second derivatives of the objective and each constraints at x against centered finite differences. This check does not rely on exactness of the first derivatives, only on objective and constraint values.

The sgn arguments refers to the formulation of the Lagrangian in the problem. It should have a positive value if the Lagrangian is formulated as

\[L(x,y) = f(x) + \sum_j yⱼ cⱼ(x),\]

and a negative value if the Lagrangian is formulated as

\[L(x,y) = f(x) - \sum_j yⱼ cⱼ(x).\]

Only the sign of sgn is important.

This function returns a dictionary indexed by functions. The 0-th function is the objective while the k-th function (for k > 0) is the k-th constraint. The values of the dictionary are dictionaries indexed by tuples (i, j) such that the relative error in the second derivative ∂²fₖ/∂xᵢ∂xⱼ exceeds rtol.

NLPModelsTest.hessian_check_from_grad — Method

hessian_check_from_grad(nlp; x=nlp.meta.x0, atol=1e-6, rtol=1e-4, sgn=1)

Check the second derivatives of the objective and each constraints at x against centered finite differences. This check assumes exactness of the first derivatives.

The sgn arguments refers to the formulation of the Lagrangian in the problem. It should have a positive value if the Lagrangian is formulated as

\[L(x,y) = f(x) + \sum_j yⱼ cⱼ(x),\]

and a negative value if the Lagrangian is formulated as

\[L(x,y) = f(x) - \sum_j yⱼ cⱼ(x).\]

Only the sign of sgn is important.

This function returns a dictionary indexed by functions. The 0-th function is the objective while the k-th function (for k > 0) is the k-th constraint. The values of the dictionary are dictionaries indexed by tuples (i, j) such that the relative error in the second derivative ∂²fₖ/∂xᵢ∂xⱼ exceeds rtol.

NLPModelsTest.jacobian_check — Method

jacobian_check(nlp; x=nlp.meta.x0, atol=1e-6, rtol=1e-4)

Check the first derivatives of the constraints at x against centered finite differences.

This function returns a dictionary indexed by (j, i) tuples such that the relative error in the i-th partial derivative of the j-th constraint exceeds rtol.

NLPModelsTest.multiple_precision_nlp — Method

multiple_precision_nlp(nlp_from_T; precisions=[...], exclude = [ghjvprod])

Check that the NLP API functions output type are the same as the input. In other words, make sure that the model handles multiple precisions.

The input nlp_from_T is a function that returns an nlp from a type T. The array precisions are the tested floating point types. Defaults to [Float16, Float32, Float64, BigFloat].

NLPModelsTest.multiple_precision_nls — Method

multiple_precision_nls(nls_from_T; precisions=[...], exclude = [])

Check that the NLS API functions output type are the same as the input. In other words, make sure that the model handles multiple precisions.

The input nls_from_T is a function that returns an nls from a type T. The array precisions are the tested floating point types. Defaults to [Float16, Float32, Float64, BigFloat].

NLPModelsTest.view_subarray_nlp — Method

view_subarray_nlp(nlp; exclude = [])

Check that the API work with views, and that the results is correct.

NLPModelsTest.view_subarray_nls — Method

view_subarray_nls(nls; exclude = [])

Check that the API work with views, and that the results is correct.